requirements.txt
torch>=2.1.0,<3.0.0
transformers>=4.36.0,<5.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
tqdm>=4.66.0
psutil>=5.9.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
huggingface-hub>=0.20.0
safetensors>=0.4.0
sentencepiece>=0.1.99
bitsandbytes>=0.41.1
accelerate>=0.26.0